generate_vec_roots in lib
